+++ plugin-viewport/src/test/kotlin/com/mapbox/maps/plugin/viewport/transition/MapboxViewportTransitionFactoryTest.kt
@@ -47,7 +47,13 @@ class MapboxViewportTransitionFactoryTest {
     every { createAnimatorSet(any()) } returns animatorSet
     every { animatorSet.constrainDurationTo(any()) } returns constrainedSet
     every { screenDistanceFromMapCenterToTarget(mapCameraDelegate, any(), any()) } returns 1000.0
-    every { cameraPlugin.createCenterAnimator(any(), any()) } returns mockk()
+    every {
+      cameraPlugin.createCenterAnimator(
+        options = any(),
+        useShortestPath = any(),
+        block = any()
+      )
+    } returns mockk()
     every { cameraPlugin.createBearingAnimator(any(), any(), any()) } returns mockk()
     every { cameraPlugin.createPitchAnimator(any(), any()) } returns mockk()
     every { cameraPlugin.createZoomAnimator(any(), any()) } returns mockk()
@@ -56,12 +62,14 @@ class MapboxViewportTransitionFactoryTest {
   }
 
   @Test
-  fun `transitionFromLowZoomToHighZoom - bearing is normalized`() {
+  fun `transitionFromLowZoomToHighZoom - bearing and center is normalized`() {
     every { mapCameraDelegate.cameraState } returns mockk {
       every { bearing } returns 10.0
+      every { center } returns Point.fromLngLat(-170.0, 0.0)
     }
     val cameraOptions = CameraOptions.Builder()
       .bearing(350.0)
+      .center(Point.fromLngLat(170.0, 10.0))
       .build()
 
     val valueSlot = slot<CameraAnimatorOptions<Double>>()
@@ -75,16 +83,25 @@ class MapboxViewportTransitionFactoryTest {
 
     assertEquals(-10.0, valueSlot.captured.targets.last(), EPS)
     verify { normalizeBearing(10.0, 350.0) }
+    verify {
+      cameraPlugin.createCenterAnimator(
+        options = any(),
+        useShortestPath = true,
+        block = any()
+      )
+    }
   }
 
   @Test
-  fun `transitionFromHighZoomToLowZoom - bearing is normalized`() {
+  fun `transitionFromHighZoomToLowZoom - bearing and center is normalized`() {
     every { mapCameraDelegate.cameraState } returns mockk {
       every { bearing } returns 10.0
+      every { center } returns Point.fromLngLat(-170.0, 0.0)
     }
 
     val cameraOptions = CameraOptions.Builder()
       .bearing(350.0)
+      .center(Point.fromLngLat(170.0, 10.0))
       .build()
 
     val valueSlot = slot<CameraAnimatorOptions<Double>>()
@@ -98,15 +115,24 @@ class MapboxViewportTransitionFactoryTest {
 
     assertEquals(-10.0, valueSlot.captured.targets.last(), EPS)
     verify { normalizeBearing(10.0, 350.0) }
+    verify {
+      cameraPlugin.createCenterAnimator(
+        options = any(),
+        useShortestPath = true,
+        block = any()
+      )
+    }
   }
 
   @Test
-  fun `transitionLinear - bearing is normalized`() {
+  fun `transitionLinear - bearing and center is normalized`() {
     every { mapCameraDelegate.cameraState } returns mockk {
       every { bearing } returns 10.0
+      every { center } returns Point.fromLngLat(-170.0, 0.0)
     }
     val cameraOptions = CameraOptions.Builder()
       .bearing(350.0)
+      .center(Point.fromLngLat(170.0, 10.0))
       .build()
 
     val valueSlot = slot<CameraAnimatorOptions<Double>>()
@@ -117,6 +143,13 @@ class MapboxViewportTransitionFactoryTest {
 
     assertEquals(-10.0, valueSlot.captured.targets.last(), EPS)
     verify { normalizeBearing(10.0, 350.0) }
+    verify {
+      cameraPlugin.createCenterAnimator(
+        options = any(),
+        useShortestPath = true,
+        block = any()
+      )
+    }
   }
 
   @Test